Moved models into a new directory
