+++ Checkout.Tests/CheckoutTests.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Checkout.Exceptions;
 using Checkout.Interfaces;
+using Checkout.Models;
 using Moq;
 using NUnit.Framework;
