+++ Checkout.Tests/DiscountServiceTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Checkout.Interfaces;
+using Checkout.Models;
 using Moq;
 using NUnit.Framework;
