+++ Checkout/Interfaces/IDiscountService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Checkout.Models;
 
 namespace Checkout.Interfaces
 {
