+++ Checkout/Models/BasketItem.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Checkout.Interfaces;
 
-namespace Checkout
+namespace Checkout.Models
 {
     /// <summary>
     /// Represents a collection of <see cref="Product"/> ready for checkout
